Isolated clock module(s)
