+++ packages/lib/src/components/internal/Calendar/calendar/facade/Calendar.ts
@@ -43,7 +43,7 @@ import { isString } from '@src/utils/validator-utils';
 import watchable from '@src/utils/watchable';
 import { Watchable, WatchableFactory, WatchCallable, WatchCallback } from '@src/utils/watchable/types';
 import { MonthFrame, TimeFrame, YearFrame } from '../timeframe';
-import { today } from '../clock';
+import { today } from '../../clock';
 import {
     CalendarConfig,
     CalendarDayOfWeekData,
