+++ packages/lib/src/components/internal/Calendar/clock/today.ts
@@ -1,7 +1,7 @@
 import $restamper, { RestamperWithTimezone, systemToTimezone, timezoneToSystem } from '@src/core/Localization/datetime/restamper';
 import $watchable from '@src/utils/watchable';
 import { struct } from '@src/utils/common';
-import { Today } from '../types';
+import { Today } from './types';
 import clock from './clock';
 
 const today = (() => {
